Get rid of unneeded symbolic link
